[조세웅] - 댓글 삭제 기능 완료 #10

+++ src/main/java/com/concertfinder/concertfinder/comment/repository/CommentRepository.java
@@ -12,5 +12,5 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     public Page<Comment> findAllByPostId(long postId, Pageable pageable);
     
-    public long countByPostId(long postId);
+    public int countByPostId(long postId);
 }
